ac_ball link

+++ include/libc64/qrand.h
@@ -6,6 +6,6 @@
 u32 qrand(void);
 void sqrand(u32);
 f32 fqrand(void);
-f64 fqrand2(void);
+f32 fqrand2(void);
 
 #endif
